loading: propose simplifications of code and docs
Attempt a simplification of the module-search logic.
Previously, it was mostly using 3VL (UUID found, UUID not-found,
environment not-found). However, it turns out that it can also
be handled via nullable (is the Pkg identified or not).

Also propose some simplifications of the wording and layout of the code-loading docs.