code show: thread debuginfo verbosity parameter throughout APIs
These levels are intended to roughly correspond approximately to compiler options -g0 (none) through -g2 (including local variables).
Currently we do not implement the ability to represent local variable information, so the default is to only show line number information
suitable for identifying the source line of a given statement (but not the types).